feat: added HomeLayout.vue and LandingPage.vue
